rulearith: biggestNonMultCoeff u32 with 3 truncation-time mirrors; spacebase byteToAddress unsigned division (CR24 R1/R2 on union 781046c2)
Union verification (781046c2 = FM 26ffb3c2 x master 9458a61b) plus the two
CR24-queued latent fixes. The three "Type propagation algorithm not settling"
warnings FM saw on base 06034903 are GONE on the union, and the cure is real
convergence, not hiding: instrumented InferTypes counters for
file2string.part.0 / getparameter.constprop.0 / match_url settle at max
localcount 3/5/5 (< 7); gp fullloop stages 371 == oracle (FM branch had 587),
myprogress 402 == oracle; parseconfig / next_url / match_url projections
MATCH; first divergences moved ord209 -> ord351 (gp) and ord173 -> ord399
(myprogress); curl 2145/0/0, httpd 2057/0/0, --func getparameter 532,
--func myprogress 64; unit tests green outside the known flaky family.
Curing channel = master range 06034903..9458a61b (live ScopeLocal channel
4f4c9d78 + 4256a1a7, input prototype 13f080fd/f8ee7548). Full gate numbers in
the lane report.

R1: oracle field is `uint4 biggestNonMultCoeff` (ruleaction.hh:54) and the
hasMatchingSubType formal is `uint4 coeff` (ruleaction.cc:6064) - no
truncation happens at the call. Rust field u64 -> u32 with the three sites
mirrored at their exact truncation points: ruleaction.cc:6146 casts to uint4
BEFORE the comparison (a wrapped-to-zero vncoeff can lose the race);
ruleaction.cc:6158-6159 and 6210-6211 compare `treeCoeff` (uint8) at full
width against the promoted uint4 field, then truncate at the store. The
previous in-code claim of a "uint8 field" was wrong and is corrected.

R2: `AddrSpace::byteToAddress(uintb val, uint4 ws) { return val/ws; }`
(space.hh:522-524) is UNSIGNED division. The three `off` sites in
TypeSpacebase (get_sub_type, get_sub_type_in_map,
nearest_arrayed_component_forward_in_map) divided as signed i64 and cast;
they now divide the two's-complement bit pattern like the oracle. The
wordsize>1 oracle fixture gap is registered as RULEARITH-SPACEBASE-USDIV-0001
(P3 latent).

Both changes are corpus-neutral (ws = 1 identity; coefficients < 2^32):
curl and httpd E2E outputs byte-identical before/after this commit.

## Alignment Evidence
Ghidra: ruleaction.hh:54 `uint4 biggestNonMultCoeff;`
Ghidra: ruleaction.cc:6146 `uint4 vncoeff = (sval < 0) ? (uint4)-sval : (uint4)sval;`
Ghidra: ruleaction.cc:6158-6159 `if (treeCoeff > biggestNonMultCoeff) biggestNonMultCoeff = treeCoeff;` (treeCoeff uint8 per ruleaction.hh:70-72)
Ghidra: space.hh:522-524 `uintb AddrSpace::byteToAddress(uintb val, uint4 ws) { return val/ws; }`
  关键决定性语义（四类，逐条核对）:
  - 引用/输出参数: 全部为字段/按值形参语义; hasMatchingSubType(uint4 coeff) 按值; byteToAddress(uintb val) 按值; 无输出参数.
  - 循环边界/遍历顺序: 不涉循环; 三个赋值位点顺序固定 — 6146 先转换后比较, 6158-6159 与 6210-6211 先比较(64 位宽)后截断存储.
  - 计数器/累加器: biggestNonMultCoeff 是运行最大值累加器, 存储域 uint4; 6146 位点截断发生在竞争之前 (wrapped 值可为 0 落选), 6158/6210 位点竞争在 uint8 宽度、存储时截断; 消费点 6271 (!=0) 与 6290/6304 读已截断存储值.
  - 排序/比较键: `>` 严格大于才覆盖 (平局保持先见者); byteToAddress 是补码位型上的无符号截断除法 val/ws.
Rugra: src/ruleaction.rs:17971 AddTreeState.biggest_non_mult_coeff (u32)
Rugra: src/ruleaction.rs:18223 check_mult_term vncoeff mirror (as u32 before compare)
Rugra: src/ruleaction.rs:18240+18304 compare `u64::from(field) < tree_coeff` then store `as u32`
Rugra: src/ruleaction.rs:18403 has_matching_sub_type (array_hint: u32, no call-site truncation)
Rugra: src/type_system/datatype.rs:4357+4464+4500 `(off as u64).wrapping_div(wordsize as u64)`
  - 引用参数: u32 字段与按值形参; 调用点传 u32 字段零截断 = oracle.
  - 遍历顺序: 三位点与 oracle 一一对应; site1 wrapping_neg 后 as u32 再比较; site2/3 u64 宽比较后 as u32 存储.
  - 计数器: 字段 u32 存储域 = oracle uint4; `!=0` 与吸附消费读 u32 存储值.
  - 排序键: `<` 严格小于覆盖镜像 oracle `>`; 无符号按位型除法镜像 uintb val/ws.
四类决定性语义核对: [x]引用参数 [x]遍历顺序 [x]计数器 [x]排序键

## Differential
curl E2E vs ghidra_curl_1204.c: 2145/0/0 (skeleton/defects/numbering), output byte-identical to the pre-change union state; httpd E2E vs ghidra_httpd_1204.c: 2057/0/0, byte-identical. Projections: parseconfig MATCH, next_url MATCH, match_url MATCH, gp first divergence ord351 (held), myprogress ord399 (held). Zero defects/numbering issues; the two behavior changes are latent-only (R1 needs |coeff| >= 2^32, R2 needs wordsize > 1; neither reachable in the current corpus). wordsize>1 fixture gap tracked as RULEARITH-SPACEBASE-USDIV-0001.

## Cross-Review
CR25 requested for this commit (ruleaction.rs Rule-domain whitelist + datatype.rs spacebase query family): audit the union decisions recorded in TODO_BOARD row RULEARITH-POSTADSORB-CONVERGE-0001 plus the R1/R2 registrations. Materials: /dev/shm/rugra-reports/LANE_FW2_POSTADSORB_2026-09-24.md and /dev/shm/rugra-tests/sb-postadsorb/.

+++ src/type_system/datatype.rs
@@ -4352,7 +4352,9 @@ impl TypeSpacebase {
         let wordsize = self.spaceid.map(|s| s.word_size()).unwrap_or(1).max(1) as i64;
         // AddrSpace::byteToAddress(off, wordsize) (space.hh:523) = off / ws
         // (FM direction fix: mul→div; ws=1 makes both identity here).
-        let addr_off = off.wrapping_div(wordsize) as u64;
+        // Unsigned uintb division: negative stack offsets divide on the
+        // two's-complement bit pattern like the oracle, not as i64.
+        let addr_off = (off as u64).wrapping_div(wordsize as u64);
         match self.get_map() {
             None => (
                 Some(Arc::new(Datatype::Base(TypeBase::new(
@@ -4456,8 +4458,10 @@ impl TypeSpacebase {
     /// like the snapshot form.
     pub fn get_sub_type_in_map(&self, map: &SpacebaseMap<'_>, off: i64) -> (Option<Arc<Datatype>>, i64) {
         let wordsize = self.spaceid.map(|s| s.word_size()).unwrap_or(1).max(1) as i64;
-        // AddrSpace::byteToAddress(off, wordsize) = off / ws (space.hh:523).
-        let addr_off = off.wrapping_div(wordsize) as u64;
+        // AddrSpace::byteToAddress(off, wordsize) = off / ws (space.hh:523),
+        // unsigned uintb division on the bit pattern (negative offsets keep
+        // the oracle's huge-positive quotient at ws > 1; ws = 1 identity).
+        let addr_off = (off as u64).wrapping_div(wordsize as u64);
         match self.query_container_in_map(map, addr_off) {
             Some(hit) => {
                 // newoff = (addr - entry.addr) + entry.offset (type.cc:2967).
@@ -4491,8 +4495,9 @@ impl TypeSpacebase {
     ) -> ArrayedComponent {
         let wordsize = self.spaceid.map(|s| s.word_size()).unwrap_or(1).max(1) as i64;
         // byteToAddress(off, ws) = off / ws (space.hh:523); resolveConstant
-        // is modelled as the identity mapping into the space.
-        let addr = off.wrapping_div(wordsize) as u64;
+        // is modelled as the identity mapping into the space. Unsigned
+        // uintb division (see get_sub_type_in_map).
+        let addr = (off as u64).wrapping_div(wordsize as u64);
         // type.cc:2984-2985 — no container, or a partial piece
         // (getOffset() != 0): probe 32 address units ahead.
         let first = match self.query_container_in_map(map, addr) {
